refactor: permission presenters
for the single permission types where many "similar" tables exist,
make base presenters (for inheritance)

for sharing of concerns (between types of MediaResources and between types of Permissions for a single Resource),
only use Modules (for including)
