+++ app/presenters/presenters/collections/permissions/collection_group_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module Collections
     module Permissions
       class CollectionGroupPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceGroupPermission
+        Presenters::Shared::Resource::ResourceGroupPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate :edit_metadata_and_relations, to: :@app_resource
       end
