+++ app/presenters/presenters/collections/permissions/collection_public_permission.rb
@@ -2,8 +2,9 @@ module Presenters
   module Collections
     module Permissions
       class CollectionPublicPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourcePublicPermission
+        Presenters::Shared::Resource::ResourcePublicPermission
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
       end
     end
   end
