+++ app/presenters/presenters/collections/permissions/collection_user_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module Collections
     module Permissions
       class CollectionUserPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceUserPermission
+        Presenters::Shared::Resource::ResourceUserPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate :edit_metadata_and_relations, to: :@app_resource
       end
