+++ app/presenters/presenters/filter_sets/permissions/filter_set_api_client_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module FilterSets
     module Permissions
       class FilterSetApiClientPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceApiClientPermission
+        Presenters::Shared::Resource::ResourceApiClientPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate :edit_metadata_and_filter, to: :@app_resource
       end
