+++ app/presenters/presenters/filter_sets/permissions/filter_set_group_permission.rb
@@ -2,8 +2,9 @@ module Presenters
   module FilterSets
     module Permissions
       class FilterSetGroupPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceGroupPermission
+        Presenters::Shared::Resource::ResourceGroupPermission
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
       end
     end
   end
