+++ app/presenters/presenters/filter_sets/permissions/filter_set_user_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module FilterSets
     module Permissions
       class FilterSetUserPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceUserPermission
+        Presenters::Shared::Resource::ResourceUserPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate_to_app_resource :edit_metadata_and_filter
       end
