+++ app/presenters/presenters/media_entries/permissions/media_entry_api_client_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module MediaEntries
     module Permissions
       class MediaEntryApiClientPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceApiClientPermission
+        Presenters::Shared::Resource::ResourceApiClientPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate :get_full_size, to: :@app_resource
       end
