+++ app/presenters/presenters/media_entries/permissions/media_entry_public_permission.rb
@@ -2,8 +2,10 @@ module Presenters
   module MediaEntries
     module Permissions
       class MediaEntryPublicPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourcePublicPermission
+        Presenters::Shared::Resource::ResourcePublicPermission
+
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
 
         delegate :get_full_size, to: :@app_resource
       end
