+++ app/presenters/presenters/media_entries/permissions/media_entry_user_permission.rb
@@ -2,14 +2,13 @@ module Presenters
   module MediaEntries
     module Permissions
       class MediaEntryUserPermission < \
-        Presenters::Shared::MediaResource::Permissions::\
-          MediaResourceUserPermission
+        Presenters::Shared::MediaResource::Permissions::MediaResourceUserPermission
 
-        include Presenters::\
-                MediaEntries::\
-                Permissions::\
-                Modules::\
-                MediaEntryCommonPermissions
+        include Presenters::Shared::MediaResource::\
+          Permissions::MediaResourceCommonPermissions
+
+        delegate :get_full_size, to: :@app_resource
+        delegate :edit_metadata, to: :@app_resource
       end
     end
   end
