--- app/presenters/presenters/shared/media_resource/permissions/media_resource_api_client_permission.rb
@@ -1,21 +1,0 @@
-module Presenters
-  module Shared
-    module MediaResource
-      module Permissions
-        class MediaResourceApiClientPermission < MediaResourceCommonPermission
-
-          def subject # TODO: ApiClientIndexPresenter
-            a = @app_resource.api_client
-            Pojo.new(
-              uuid: a.id,
-              login: a.login,
-              description: a.description,
-              _type: 'ApiClientIndex'
-            )
-          end
-
-        end
-      end
-    end
-  end
-end
