--- app/presenters/presenters/shared/media_resource/permissions/media_resource_group_permission.rb
@@ -1,15 +1,0 @@
-module Presenters
-  module Shared
-    module MediaResource
-      module Permissions
-        class MediaResourceGroupPermission < MediaResourceCommonPermission
-
-          def subject
-            Presenters::Groups::GroupIndex.new(@app_resource.group)
-          end
-
-        end
-      end
-    end
-  end
-end
