--- app/presenters/presenters/shared/media_resource/permissions/media_resource_public_permission.rb
@@ -1,10 +1,0 @@
-module Presenters
-  module Shared
-    module MediaResource
-      module Permissions
-        class MediaResourcePublicPermission < MediaResourceCommonPermission
-        end
-      end
-    end
-  end
-end
