+++ app/presenters/presenters/shared/media_resource/permissions/media_resource_user_permission.rb
@@ -2,16 +2,11 @@ module Presenters
   module Shared
     module MediaResource
       module Permissions
-        class MediaResourceUserPermission < MediaResourceCommonPermission
-          def initialize(app_resource)
-            @app_resource = app_resource
-          end
-
-          def subject
-            Presenters::Users::UserIndex.new(@app_resource.user)
-          end
+        class MediaResourceUserPermission < \
+          Presenters::Shared::Resource::ResourceUserPermission
 
           delegate :edit_permissions, to: :@app_resource
+
         end
       end
     end
