+++ app/presenters/presenters/shared/resource/resource_api_client_permission.rb
@@ -1,0 +1,18 @@
+module Presenters
+  module Shared
+    module Resource
+      class ResourceApiClientPermission < Presenters::Shared::AppResource
+
+        def subject # TODO: ApiClientIndexPresenter
+          a = @app_resource.api_client
+          Pojo.new(
+            uuid: a.id,
+            login: a.login,
+            description: a.description,
+            _type: 'ApiClientIndex'
+          )
+        end
+      end
+    end
+  end
+end
