+++ app/presenters/presenters/shared/resource/resource_group_permission.rb
@@ -1,0 +1,13 @@
+module Presenters
+  module Shared
+    module Resource
+      class ResourceGroupPermission < Presenters::Shared::AppResource
+
+        def subject
+          Presenters::Groups::GroupIndex.new(@app_resource.group)
+        end
+
+      end
+    end
+  end
+end
