+++ app/presenters/presenters/shared/resource/resource_public_permission.rb
@@ -1,0 +1,8 @@
+module Presenters
+  module Shared
+    module Resource
+      class ResourcePublicPermission < Presenters::Shared::AppResource
+      end
+    end
+  end
+end
