+++ app/presenters/presenters/shared/resource/resource_user_permission.rb
@@ -1,0 +1,13 @@
+module Presenters
+  module Shared
+    module Resource
+      class ResourceUserPermission < Presenters::Shared::AppResource
+
+        def subject
+          Presenters::Users::UserIndex.new(@app_resource.user)
+        end
+
+      end
+    end
+  end
+end
